+++ modules-services2/service-templates/src/main/java/org/openprovenance/prov/template/service/ProvapiApplication.java
@@ -30,10 +30,10 @@ import org.openprovenance.prov.service.translation.TranslationService;
 import jakarta.ws.rs.ApplicationPath;
 import jakarta.ws.rs.core.Application;
 import org.openprovenance.prov.service.translation.storage.StorageSetup;
-import org.openprovenance.prov.template.service.readers.JsonOrCsvMessageBodyReader;
-import org.openprovenance.prov.template.service.readers.SearchConfigMessageBodyReader;
-import org.openprovenance.prov.template.service.readers.TableKeyListMessageBodyReader;
-import org.openprovenance.prov.template.service.readers.TemplatesVizConfigMessageBodyReader;
+import org.openprovenance.prov.service.core.readers.JsonOrCsvMessageBodyReader;
+import org.openprovenance.prov.service.core.readers.SearchConfigMessageBodyReader;
+import org.openprovenance.prov.service.core.readers.TableKeyListMessageBodyReader;
+import org.openprovenance.prov.service.core.readers.TemplatesVizConfigMessageBodyReader;
 
 
 import java.util.HashSet;