+++ modules-services2/service-templates/src/main/java/org/openprovenance/prov/template/service/readers/TemplatesVizConfig.java
@@ -1,8 +1,0 @@
-package org.openprovenance.prov.template.service.readers;
-
-public class TemplatesVizConfig {
-    public Integer id;
-    public String template;
-    public String property;
-    public String style;
-}