+++ RELEASES.md
@@ -40,6 +40,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: changed Locations to better manager packages of generated code
 * prov-template-compiler: added generated method isCompositeOf in Builder
 * prov-template-archetype: support for compulsory properties in forms.html for composites
+* service-templates: refactored to rely on service-templates-core
+
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java